seis-colores/scripts/requirements.txt
joblib==1.4.2
numpy==2.2.4
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.2
threadpoolctl==3.6.0
